backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
bcrypt==4.0.1
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib[bcrypt]==1.7.4
tzdata>=2024.2
motor==3.3.1
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.9
jq>=1.6.0
typer>=0.9.0
websockets>=12.0

# Enhanced backend dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0
psutil>=5.9.0
pyyaml>=6.0.0
asyncpg>=0.29.0  # For PostgreSQL async support
aiomysql>=0.2.0  # For MySQL async support
redis>=5.0.0  # For caching and sessions
celery>=5.3.0  # For background tasks
flower>=2.0.0  # Celery monitoring
python-json-logger>=2.0.0  # Structured JSON logging
slowapi>=0.1.9  # Rate limiting
pydantic-settings>=2.1.0  # Settings management
itsdangerous>=2.1.0  # Secure token generation
blinker>=1.7.0  # Signal system for events
marshmallow>=3.20.0  # Data serialization
apispec>=6.3.0  # API documentation
gspread>=5.12.0
google-auth-oauthlib>=1.2.0
pandas>=2.2.0
